source/requirements.txt
Flask==2.3.3
gunicorn==23.0.0
google-cloud-logging==3.6.0
google-cloud-firestore==2.12.0
google-cloud-storage==2.10.0
google-cloud-secret-manager==2.16.4
google-api-python-client==2.101.0
google-auth==2.23.1
google-auth-oauthlib==1.1.0
google-cloud-translate==3.12.0
psycopg2==2.9.10
alembic==1.15.2
